adding file used for data wrangling
Note: For percentage values for which the data was suppressed, I replaced them with zeroes and then noted which measures they belonged to such that, when we divide for the average at the end, we don't take into account that type that was zeroed. For example, for csv 9, the "Distant" measure had suppressed data. I changed this to a 0, and then divided by 1 fewer when taking the average at the end.

The following where edited:
9. Distant
12. Regional, unstaged
17. Regional, unstaged